### 2025-01-09 #7640 @warrickball
Hi @lgrcia — Thanks for your submission to JOSS. Owing to the relatively small size of the codebase, I'm going to flag this for editorial review on whether it meets [our requirements for substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This process should take about a week or two while we come to a consensus. Thanks for your patience!
