### 2025-01-24 aba6178
fix: add z coordinate
