+++ spotter/light_curves.py
@@ -71,7 +71,12 @@ def light_curve(star: Star, time: ArrayLike) -> ArrayLike:
 
 
 def transit_light_curve(
-    star: Star, x: float = 0.0, y: float = 0.0, r: float = 0.0, time: float = 0.0
+    star: Star,
+    x: float = 0.0,
+    y: float = 0.0,
+    z: float = 0.0,
+    r: float = 0.0,
+    time: float = 0.0,
 ):
     """Light curve of a transited Star. The x-axis cross the star in the horizontal direction (→),
     and the y-axis cross the star in the vertical up direction (↑).
@@ -93,4 +98,4 @@ def transit_light_curve(
     ArrayLike
         Light curve array.
     """
-    return light_curve(transited_star(star, y, x, r), star.phase(time))
+    return light_curve(transited_star(star, y, x, z, r), star.phase(time))
