### 2025-01-31 #7640 @warrickball
Hi again @lgrcia, after enquiring among our editorial team I've decided that `spotter` is a substantial contribution, despite the smallish codebase, so we'll continue our JOSS submission process. One point that was raised in our discussion is that the additional value of `spotter` (especially over e.g. `starry-process`) should be made stronger, perhaps by adding something like an example of a starspot being confused for a transit or transit parameters being biased because of a spotty stellar surface (either of which I believe could be generated using `spotter`).

That said, all the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!
