### 2025-01-31 34ed37a
fix: transited star
