### 2025-02-03 #7640 @lgrcia
Thanks for the update @warrickball!

Indeed, this is one of the main application of `spotter` right now. Would your suggestion be included in the JOSS paper or in the documentation as a use case?
