### 2025-02-03 cf51728
add convenience properties and methods
