### 2025-02-03 b6a7abc
feat: switch to uv + spotter name on pypi!
