### 2025-02-03 455a230
fix: None time in light curves
