### 2025-02-03 ddd7d47
ci: update to uv
