+++ .github/workflows/ci.yml
@@ -39,20 +39,16 @@ jobs:
           fetch-depth: 0
           submodules: true
 
-      - name: "Init: Python"
-        uses: actions/setup-python@v5
+      - name: Install uv and set the python version
+        uses: astral-sh/setup-uv@v5
         with:
           python-version: ${{ matrix.python-version }}
 
-      - name: "Install: dependencies"
-        run: |
-          python -m pip install -U pip
-          python -m pip install -U nox
+      - name: Install the project
+        run: uv sync --all-extras --dev
 
-      - name: "Tests: run"
-        run: |
-          python -m nox --non-interactive --error-on-missing-interpreter \
-            --session "${{matrix.session}}-${{matrix.python-version}}"
+      - name: Run tests
+        run: uv run pytest tests
 
   build:
     runs-on: ubuntu-latest
@@ -79,7 +75,7 @@ jobs:
   publish:
     environment:
       name: pypi
-      url: https://pypi.org/p/starspotter
+      url: https://pypi.org/p/spotter
     permissions:
       id-token: write
     needs: [tests, build]
