+++ .github/workflows/ci.yml
@@ -14,6 +14,7 @@ on:
 jobs:
   tests:
     runs-on: ${{ matrix.os }}
+    needs: ["build"]
     strategy:
       fail-fast: false
       matrix:
@@ -45,7 +46,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Install the project
-        run: uv sync --all-extras --dev
+        run: uv sync --dev
 
       - name: Run tests
         run: uv run pytest tests
@@ -56,18 +57,23 @@ jobs:
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
-      - uses: actions/setup-python@v5
-        name: Install Python
+
+      - name: Install uv and set the python version
+        uses: astral-sh/setup-uv@v5
         with:
           python-version: "3.11"
-      - name: Install dependencies
-        run: |
-          python -m pip install -U pip
-          python -m pip install -U build twine
-      - name: Build the distribution
-        run: python -m build .
+
+      - name: Install the project
+        run: uv sync --dev
+
+      - name: Build the project
+        run: uv build
+
       - name: Check the distribution
-        run: python -m twine check --strict dist/*
+        run: |
+          uv tool install twine
+          uvx twine check --strict dist/*
+
       - uses: actions/upload-artifact@v4
         with:
           path: dist/*
