### 2025-02-03 17bec3d
fix: add uv extra
