+++ .github/workflows/ci.yml
@@ -46,7 +46,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Install the project
-        run: uv sync --dev
+        run: uv sync --extra test --extra comparison
 
       - name: Run tests
         run: uv run pytest tests
