### 2025-02-03 f39d0ad
fix: update to jaxoplanet 0.0.3
