### 2025-02-04 #7640 @warrickball
I would suggest including it in the JOSS paper, and then using that example as the source to demonstrate the speedup in the GPU version compared to the CPU version. The idea here is to quickly demonstrate why `spotter` is a useful contribution to existing ecosystem of software tools in this area.

We don't doubt that `spotter` is useful but we want the paper to demonstrate as directly as possible _why_ it's useful.
