### 2025-02-06 38b9800
feat: revamped some of spotter tutorials
