+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "spotter"
 version = "0.0.3"
-description = "Forward models of fluxes and spectra time-series of non-uniform stars."
+description = "Approximate forward models of fluxes and spectra time-series of non-uniform stars."
 authors = [{ name = "Lionel Garcia" }, { name = "Benjamin Rackham" }]
 license = "MIT"
 readme = "readme.md"
@@ -9,7 +9,7 @@ requires-python = ">=3.10"
 dependencies = ["numpy", "healpy", "jax", "jaxlib", "equinox", "tinygp"]
 
 [project.optional-dependencies]
-dev = ["black", "pytest", "nox"]
+dev = ["black", "pytest", "nox", "ipykernel", "matplotlib"]
 test = ["pytest", "pytest-xdist"]
 comparison = ["jaxoplanet>=0.0.3"]
 docs = [
