### 2025-02-06 0e2be4d
fix: remove white contour and adjust show xlim-ylim
