### 2025-02-12 b9ca446
feat: radial velocity
