+++ spotter/core.py
@@ -116,7 +116,7 @@ def soft_spot(N, latitude, longitude, radius):
     return profile / jnp.max(profile)
 
 
-def render(y, inc=None, u=None, phase=0.0, obl=0.0):
+def render(y, inc=None, u=None, phase=0.0, obl=0.0, xsize=800):
     import matplotlib.pyplot as plt
 
     X = vec(y)
@@ -126,7 +126,9 @@ def render(y, inc=None, u=None, phase=0.0, obl=0.0):
         rot=[phase, np.pi / 2 - inc or 0.0, obl or 0.0], deg=False
     ).rotate_map_pixel(y * limb_darkening)
 
-    projected_map = hp.orthview(rotated, half_sky=True, return_projected_map=True)
+    projected_map = hp.orthview(
+        rotated, half_sky=True, return_projected_map=True, xsize=xsize
+    )
     plt.close()
 
     return projected_map
@@ -169,14 +171,21 @@ def transit_chord(N, x, r, inc=None):
     return jnp.abs(_x - x) < r
 
 
-def doppler_shift(theta, phi, period, radius, phase):
+def radial_velocity(N, theta, phi, period, radius, phase, inc, intensity):
     period_s = period * 24 * 60 * 60  # convert days to seconds
     omega = jnp.pi * 2 / period_s  # angular velocity
     radius_m = radius * 695700000.0  # convert solar radii to meters
-    c = 299792458.0
     sin_phi = np.sin(phi)  # numpy here! as phi is static
-    radial_velocity = radius_m * omega * jnp.sin(theta - phase) * sin_phi
-    shift = radial_velocity / c
+    mask, projected, limb = mask_projected_limb(vec(N), phase, inc=inc)
+    limb_geometry = projected * mask * limb
+    rv = radius_m * omega * jnp.sin(theta - phase) * sin_phi
+    return jnp.sum(rv * limb_geometry * intensity, 0)
+
+
+def doppler_shift(theta, phi, period, radius, phase):
+    rv = radial_velocity(theta, phi, period, radius, phase)
+    c = 299792458.0
+    shift = rv / c
     return shift
 
 
