### 2025-02-19 6370f8c
fix: radial velocity evaluation (design matrix multiplied twice)
