+++ spotter/star.py
@@ -225,7 +225,7 @@ def show(star: Star, phase: ArrayLike = 0.0, ax=None, xsize=800, **kwargs):
         star.inc if star.inc is not None else np.pi / 2,
         star.obl if star.obl is not None else 0.0,
         star.u[0] if star.u is not None else None,
-        phase,
+        phase=phase,
         ax=ax,
         xsize=xsize,
         **kwargs,
