### 2025-03-17 2c4ffa9
feat: rv test and tutorial
