### 2025-05-29 #8305 @editorialbot
[REVIEW]: spotter: Hardware-Accelerated Forward Models of Pixelated Stars
**Submitting author:** @lgrcia (<a href="http://orcid.org/0000-0002-4296-2246">Lionel Garcia</a>)
**Repository:** https://github.com/lgrcia/spotter
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.0.11
