### 2025-06-03 lgrcia/spotter#30 @nenasedk
[Documentation request] Contributions
Could you please add a brief statement to the documentation describing how you would like to include contributions to `spotter`? Even just something simple about forking the project and creating a pull request would be useful, but the more detailed the better, particularly for less experienced users.

At the same time, it would be good to include a brief description of how to run the tests, and that it's possible to install the optional dependencies. As a note for mac users, it's necessary to install the optional dependencies as `pip install spotter\[test\]`, since the `zsh` terminal has some issues with square brackets.
