### 2025-06-03 lgrcia/spotter#31 @nenasedk
Test suite dependencies
When running the test suite, a significant portion of the tests are skipped when using the standard test dependencies. `jaxoplanet` should be included as a test dependency to ensure full test coverage.
