### 2025-06-06 867adf1
feat: add flux kernel + docs with inference
