### 2025-06-06 1b4e25c
fix: add jaxoplanet in plain test dependencies and tweak CI
