+++ .github/workflows/ci.yml
@@ -17,20 +17,10 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest"]
+        os: ["ubuntu-latest", "macos-latest"]
         python-version: ["3.10", "3.11"]
         session:
           - "test"
-        include:
-          - os: "ubuntu-latest"
-            python-version: "3.11"
-            session: "test"
-          - os: "macos-latest"
-            python-version: "3.11"
-            session: "test"
-          - os: "ubuntu-latest"
-            python-version: "3.10"
-            session: "comparison"
 
     steps:
       - name: "Init: checkout"
@@ -45,7 +35,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Install the project
-        run: uv sync --extra test --extra comparison
+        run: uv sync --extra test
 
       - name: Run tests
         run: uv run pytest tests
