### 2025-06-06 645c9e6
docs: add flux gp to the documentation
