### 2025-06-06 c76ae21
docs: docstrings for all functions
