+++ spotter/__init__.py
@@ -1,2 +1,1 @@
-from spotter.distributions import butterfly, uniform
 from spotter.star import Star, show, video
