+++ spotter/viz.py
@@ -7,6 +7,26 @@ _DEFAULT_CMAP = "magma"
 
 
 def lon_lat_lines(n: int = 6, pts: int = 100, radius: float = 1.0):
+    """
+    Generate latitude and longitude lines on a sphere.
+
+    Parameters
+    ----------
+    n : int or tuple, optional
+        Number of latitude lines (and longitude lines if tuple).
+    pts : int, optional
+        Number of points per line.
+    radius : float, optional
+        Sphere radius.
+
+    Returns
+    -------
+    lat : ndarray
+        Latitude lines.
+    lon : ndarray
+        Longitude lines.
+    """
+
     assert isinstance(n, int) or len(n) == 2
 
     if isinstance(n, int):
@@ -43,6 +63,24 @@ def lon_lat_lines(n: int = 6, pts: int = 100, radius: float = 1.0):
 
 
 def rotation(inc, obl, theta):
+    """
+    Compute the rotation for given inclination, obliquity, and phase.
+
+    Parameters
+    ----------
+    inc : float
+        Inclination in radians.
+    obl : float
+        Obliquity in radians.
+    theta : float
+        Rotation phase in radians.
+
+    Returns
+    -------
+    R : scipy.spatial.transform.Rotation
+        Rotation object.
+    """
+
     u = [np.cos(obl), np.sin(obl), 0]
     u /= np.linalg.norm(u)
     u *= inc
@@ -55,6 +93,26 @@ def rotation(inc, obl, theta):
 
 
 def rotate_lines(lines, inc, obl, theta):
+    """
+    Rotate lines by given inclination, obliquity, and phase.
+
+    Parameters
+    ----------
+    lines : ndarray
+        Input lines.
+    inc : float
+        Inclination in radians.
+    obl : float
+        Obliquity in radians.
+    theta : float
+        Rotation phase in radians.
+
+    Returns
+    -------
+    rotated_lines : ndarray
+        Rotated lines.
+    """
+
     R = rotation(inc, obl, theta)
 
     rotated_lines = np.array([R.apply(l.T) for l in lines]).T
@@ -64,6 +122,21 @@ def rotate_lines(lines, inc, obl, theta):
 
 
 def plot_lines(lines, axis=(0, 1), ax=None, **kwargs):
+    """
+    Plot lines on a matplotlib axis.
+
+    Parameters
+    ----------
+    lines : ndarray
+        Lines to plot.
+    axis : tuple, optional
+        Axes to plot (default (0, 1)).
+    ax : matplotlib axis, optional
+        Axis to plot on.
+    **kwargs
+        Additional plot arguments.
+    """
+
     import matplotlib.pyplot as plt
 
     if ax is None:
@@ -92,6 +165,29 @@ def graticule(
     ax=None,
     **kwargs,
 ):
+    """
+    Plot a graticule (latitude/longitude grid) on a sphere.
+
+    Parameters
+    ----------
+    inc : float
+        Inclination in radians.
+    obl : float, optional
+        Obliquity in radians.
+    theta : float, optional
+        Rotation phase in radians.
+    pts : int, optional
+        Number of points per line.
+    radius : float, optional
+        Sphere radius.
+    n : int or tuple, optional
+        Number of latitude/longitude lines.
+    ax : matplotlib axis, optional
+        Axis to plot on.
+    **kwargs
+        Additional plot arguments.
+    """
+
     import matplotlib.pyplot as plt
 
     _inc = np.pi / 2 - inc
@@ -115,6 +211,29 @@ def graticule(
 
 
 def show(y, inc=np.pi / 2, obl=0.0, u=None, xsize=800, phase=0.0, ax=None, **kwargs):
+    """
+    Show a rendered map with graticule.
+
+    Parameters
+    ----------
+    y : array_like
[…]
