+++ spotter/core.py
@@ -7,16 +7,61 @@ from spotter import utils
 
 
 def distance(X, x):
+    """
+    Compute the great-circle distance between vectors X and x.
+
+    Parameters
+    ----------
+    X : array_like, shape (..., 3)
+        Array of 3D vectors.
+    x : array_like, shape (3,)
+        Single 3D vector.
+
+    Returns
+    -------
+    d : array_like
+        Great-circle distance(s) in radians.
+    """
     return jnp.arctan2(
         jnp.linalg.norm(jnp.cross(X.T, x[:, None], axis=0), axis=0), X @ x
     )
 
 
 def npix(N):
+    """
+    Return the number of HEALPix pixels for a given nside.
+
+    Parameters
+    ----------
+    N : int
+        HEALPix nside parameter.
+
+    Returns
+    -------
+    n : int
+        Number of pixels.
+    """
     return hp.nside2npix(N)
 
 
 def equator_coords(phi=None, inc=None, obl=None):
+    """
+    Compute the coordinates of the stellar equator for given orientation.
+
+    Parameters
+    ----------
+    phi : float or None, optional
+        Rotation phase in radians.
+    inc : float or None, optional
+        Inclination in radians.
+    obl : float or None, optional
+        Obliquity in radians.
+
+    Returns
+    -------
+    coords : ndarray, shape (3,)
+        Cartesian coordinates.
+    """
     # inc = None -> inc = pi/2
     if inc is None:
         si = 1.0
@@ -53,6 +98,31 @@ def equator_coords(phi=None, inc=None, obl=None):
 
 
 def mask_projected_limb(X, phase=None, inc=None, u=None, obl=None):
+    """
+    Compute mask, projected area, and limb darkening for visible pixels.
+
+    Parameters
+    ----------
+    X : array_like, shape (..., 3)
+        Cartesian coordinates of pixels.
+    phase : float, optional
+        Rotation phase in radians.
+    inc : float, optional
+        Inclination in radians.
+    u : array_like or None, optional
+        Limb darkening coefficients.
+    obl : float, optional
+        Obliquity in radians.
+
+    Returns
+    -------
+    mask : ndarray
+        Boolean mask for visible pixels.
+    projected_area : ndarray
+        Projected area for each pixel.
+    limb_darkening : ndarray
+        Limb darkening factor for each pixel.
+    """
     d = distance(X, equator_coords(phase, inc, obl))
     mask = d < jnp.pi / 2
     z = jnp.cos(d)
@@ -68,6 +138,21 @@ def mask_projected_limb(X, phase=None, inc=None, u=None, obl=None):
 
 
 def _N_or_Y_to_N_n(N_or_y):
+    """
+    Convert nside or map to (nside, npix).
+
+    Parameters
+    ----------
+    N_or_y : int or array_like
+        HEALPix nside or map.
+
+    Returns
+    -------
+    N : int
+        nside.
+    n : int
+        Number of pixels.
+    """
     if isinstance(N_or_y, int):
         n = hp.nside2npix(N_or_y)
         N = N_or_y
@@ -78,11 +163,45 @@ def _N_or_Y_to_N_n(N_or_y):
 
 
 def vec(N_or_y):
+    """
+    Return xyz coordinates for all pixels of a HEALPix map.
+
+    Parameters
+    ----------
+    N_or_y : int or array_like
+        HEALPix nside or map.
+
+    Returns
+    -------
+    xyz : ndarray, shape (npix, 3)
+        Cartesian coordinates of pixels.
+    """
     N, n = _N_or_Y_to_N_n(N_or_y)
     return np.array(hp.pix2vec(N, range(n))).T
 
 
[…]

[…]
