### 2025-06-06 a7ac354
doc: add contribution and test guides
