+++ docs/installation.md
@@ -1,5 +1,7 @@
 # Installation
 
+## Installation
+
 To install *spotter* from pypi
     
 ```bash
@@ -11,4 +13,18 @@ As *spotter* is still under development, we recommend installing the latest vers
 ```bash
 git clone https://github.com/lgrcia/spotter
 pip install -e spotter
+```
+
+## Testing
+
+To run the test suite, first install the development dependencies:
+
+```bash
+pip install "spotter[test]"
+```
+
+Then run the tests using `pytest`:
+
+```bash
+pytest tests
 ```
