### 2025-06-06 lgrcia/spotter#30 @lgrcia
Thanks for the comment! Done in #33
