### 2025-06-06 lgrcia/spotter#31 @lgrcia
I originally separated these tests to a *comparison* dependency group but I know added jaxoplanet in the main *test* group (see #32 and #33)
