### 2025-06-06 fa5be76
doc: add link to RTD doc in readme
