### 2025-06-06 #8305 @lgrcia
Hi @nenasedk and thanks for the thorough review!

For the package:

> The API documentation is somewhat lacking. There are no docstrings for most functions, and in particular it’s difficult to know what everything does. For example, what is the butterfly distribution?

I added docstrings to all functions and classes (with examples) in #33

>  The documentation needs to be mildly fleshed out in general. There should be a contribution statement, and a bit more information on how to run the tests (even just saying that the optional dependencies exist and can be installed with `pip install spotter[test]` would be helpful there. (See issues [30](https://github.com/lgrcia/spotter/issues/30) and [31](https://github.com/lgrcia/spotter/issues/31)). Likewise there should be a contact/help link somewhere in the documentation, even if it’s just a link to the GitHub issues.

I added a contribution guide and some instruction to test the project in a clean environment in #33

>   A link to the documentation should be included in the readme.

Done in [fa5be76](https://github.com/lgrcia/spotter/commit/fa5be76afccde156d57e27af8494b6bbfb61de8e)

>   One of the messages in the paper is that spotter is useful for inferring the stellar surface. An example of such an inference using e.g. `emcee` or `JAXNS` would be very helpful for others who want to incorporate it into their workflows.

Made this example in #32 and added it in #33. See [this tutorial](https://spotter.readthedocs.io/en/latest/notebooks/flux_gp/) where I use spotter and numpyro to infer the spots contrast and scale for an ensemble of light curves.

I'll provide my edits for the paper in another comment!
