### 2025-06-06 3eb9d95
doc: module level docstrings
