+++ spotter/__init__.py
@@ -1,1 +1,7 @@
+"""
+This package provides tools for modeling, visualizing, and analyzing
+non-uniform stellar surfaces using HEALPix and JAX.
+
+"""
+
 from spotter.star import Star, show, video
