### 2025-06-06 4ff4f5c
doc: add figure showing the map
