### 2025-06-06 13eabb0
feat: add pargagraph to describe the advantages of spotter over @luger:2021 and @morris:2020
