### 2025-06-06 #8305 @lgrcia
For the paper:
>  While I fully agree that `spotter` is unique and necessary, the statement of need could expand on how it is unique compared to similar tools, e.g. `starry`. Why should users use `spotter` instead?

I added a paragraph in [13eabb0](https://github.com/lgrcia/spotter/commit/13eabb059eda8f8181e9fda162e3bbab56e7f018) about the advantage of spotter over other methods/packages.

>  line 41: by high resolution, do you mean high time resolution or spectral resolution?

I meant both spatial and temporal. I highlighted that in [4ff4f5c](https://github.com/lgrcia/spotter/commit/4ff4f5c804b9beb0f856b7b088b765ff121a4271).

>  I’d recommend including a figure of a surface map in the paper, rather than relying on the documentation. This would clarify what exactly is being produced by the linear model.

I added a figure to show y, X and a complete map in [4ff4f5c](https://github.com/lgrcia/spotter/commit/4ff4f5c804b9beb0f856b7b088b765ff121a4271).

Please let me know if there is anything else I can do and thanks a lot for this review so far!
