### 2025-06-06 867ad86
fix: adapt spot crossing tuto to work with jaxoplanet >= 0.1.0
