### 2025-06-11 lgrcia/spotter#35 @tedjohnson12
[joss-review] Performance-focused example
I am really excited about this project's use of JAX, and I think it would be useful to include an example that would allow first-time users to test their own hardware acceleration against the baseline from the JOSS paper.

For example, Figure 1 of the paper compares model evaluation speed using CPU and GPU-based jaxlib implementations. An example in the docs that partially reproduces this figure (probably with just a single curve utilizing whatever implementation is currently installed) would allow first-time JAX users to 1) confirm that they are using their hardware properly, and 2) compare their own evaluation speed to your established baseline.

If such an example is not too difficult to put together, I think it would be very helpful to `spotter`'s users. Thanks!
