### 2025-06-12 0e755da
feat: add speed benchmark and more info about spotter science case / purpose
