### 2025-06-12 e36def3
fix: seeding flux test
