### 2025-06-12 efdf6f6
fix: readme links to docs
