+++ readme.md
@@ -17,10 +17,10 @@ Approximate forward models of fluxes and spectra time-series of non-uniform star
 </p>
 
 *spotter* serves as a flexible, low-level backend for scientific applications that require modeling stellar surfaces using spherical pixelation. While it is not a comprehensive simulation suite like [StarSim](https://ui.adsabs.harvard.edu/abs/2016A&A...586A.131H/abstract) or [SOAP](https://ui.adsabs.harvard.edu/abs/2014ApJ...796..132D/abstract), *spotter* provides core building blocks for custom workflows, including:
-- Modeling [surface features](https://spotter.readthedocs.io/en/latest/notebooks/introduction.ipynb) (e.g., stellar spots or faculae) with arbitrary shapes
-- Forward modeling of stellar surface [flux](https://spotter.readthedocs.io/en/latest/notebooks/multiband.ipynb), [spectra](https://spotter.readthedocs.io/en/latest/notebooks/spectral.ipynb), [radial velocities](https://spotter.readthedocs.io/en/latest/notebooks/rv.ipynb), and CCFs
-- Simulation of [observables during transits](https://spotter.readthedocs.io/en/latest/notebooks/spot_crossing.ipynb) by stellar or planetary companions
-- Gaussian process frameworks for representing [stellar surfaces](https://spotter.readthedocs.io/en/latest/notebooks/surface_gp.ipynb) and [modeling their flux time series](https://spotter.readthedocs.io/en/latest/notebooks/flux_gp.ipynb)
+- Modeling [surface features](https://spotter.readthedocs.io/en/latest/notebooks/introduction/) (e.g., stellar spots or faculae) with arbitrary shapes
+- Forward modeling of stellar surface [flux](https://spotter.readthedocs.io/en/latest/notebooks/multiband/), [spectra](https://spotter.readthedocs.io/en/latest/notebooks/spectral/), [radial velocities](https://spotter.readthedocs.io/en/latest/notebooks/rv/), and CCFs
+- Simulation of [observables during transits](https://spotter.readthedocs.io/en/latest/notebooks/spot_crossing/) by stellar or planetary companions
+- Gaussian process frameworks for representing [stellar surfaces](https://spotter.readthedocs.io/en/latest/notebooks/surface_gp/) and [modeling their flux time series](https://spotter.readthedocs.io/en/latest/notebooks/flux_gp/)
 
 *spotter* uses the [HEALPix](https://healpix.sourceforge.io/) subdivision scheme and is powered by the high-performance numerical package [JAX](https://jax.readthedocs.io/en/latest/https://spotter.readthedocs.io/en/latest/notebooks/quickstart.html), enabling its use on multiple CPUs and accelerators like GPUs and TPUs.
 
