### 2025-06-12 lgrcia/spotter#35 @lgrcia
Thanks for the issue @tedjohnson12. I've added some code to perform the speed benchmark in 0e755dab1ab57a59f615a0a70f306205d3e567ff (see [here](https://spotter.readthedocs.io/en/latest/notebooks/precision/#speed)).

@nenasedk, I did not run into specific issues depending on the JAX version yet. I'm also not sure I want to enter into hardware/jax specific issues in the documentation itself, but I am happy to set some bounds for the JAX version if anything comes up in your tests.
