### 2025-06-12 215ad5a
fic(docs): run precision offline
