### 2025-06-20 lgrcia/spotter#35 @tedjohnson12
Thanks @lgrcia! This is a really great example. I was able to get a >10x performance boost by setting `device="gpu"`, and got similar results to you with just my cpu.

As for @nenasedk's suggestion, I think it might be sufficient to just link to JAX installation documentation on your installation page. I think requiring jaxlib<=0.5.0 could be unnecessarily restrictive, so as long as you don't require >0.5.0 it should be fine.
