### 2025-06-21 lgrcia/spotter#35 @nenasedk
Thanks for updating the docs and the new example! I agree, probably unnecessary to be too restrictive on the Jax version, and unless people are regularly raising issues having trouble with installation, I think hardware specific stuff could be overkill.
