### 2025-06-23 #8305 @adonath
Thanks @nenasedk! On the GPU performance check:

@lgrcia Is there a short example that could be copied over into a Google collab for @nenasedk to evaluate the performance claims? Google collaborated allows to change the runtime to GPU / TPU.
