### 2025-06-23 #8305 @lgrcia
@adonath, I added the script [in the documentation](https://spotter.readthedocs.io/en/latest/notebooks/precision/#speed) as part of https://github.com/lgrcia/spotter/issues/35. I think @tedjohnson12 could test it on a GPU.
