### 2025-06-27 0497720
:chore: added logic to expose `spotter.__version__`
