### 2025-06-27 343a06a
fix: __version__ (#38)
