### 2025-07-01 #8305 @adonath
@nenasedk and @tedjohnson12 as far as I can tell the only remaining issue is https://github.com/lgrcia/spotter/issues/35, where @lgrcia added the example to the docs. Does one of you have time to verify the example?
