### 2025-07-01 #8305 @nenasedk
@adonath

Apologies for the delay. I've been trying to run the script that was added in the documentation, as I was recently awarded some GPU time, but have been having some issues getting everything to work. @tedjohnson12 confirmed that the script works and produced similar results to the performance claims in the paper, so the issues I'm having are on my end - @lgrcia, @tedjohnson12 have either of you ever encountered segmentation fault errors when running that script?
