### 2025-07-01 #8305 @adonath
From the editorial side the confirmation by @tedjohnson12 would be sufficient, as it would verify the performance claims in the paper. Further technical issues could be sorted out independently.

@tedjohnson12 Can you shortly confirm the script works for you and reproduces the performance claims?
