### 2025-07-01 #8305 @tedjohnson12
@adonath Yes, I can confirm the script works and I was able to verify the performance claims by switching to my GPU.
