### 2025-07-09 45303ad
Added option to choose normalize or not integ spec
