### 2025-07-10 febf58c
Changed normalize integ spec to be passed as kwarg
