### 2025-07-10 8495b74
fix: check spectrum input and shape of outputs
