+++ .github/workflows/ci.yml
@@ -1,15 +1,15 @@
 name: spotter/ci
 
 on:
-  merge_group:
-  workflow_dispatch:
-  pull_request:
-  release:
-    types: [published]
+  push:
     branches:
       - main
     tags:
       - "*"
+  pull_request:
+  merge_group:
+  release:
+    types: [published]
 
 jobs:
   tests:
@@ -70,7 +70,7 @@ jobs:
   publish:
     environment:
       name: pypi
-      url: https://pypi.org/p/spotter
+      url: https://pypi.org/p/fitsm
     permissions:
       id-token: write
     needs: [tests, build]
@@ -79,6 +79,6 @@ jobs:
     steps:
       - uses: actions/download-artifact@v4
         with:
-          name: artifact
+          name: dist
           path: dist
-      - uses: pypa/gh-action-pypi-publish@v1.8.11
+      - uses: pypa/gh-action-pypi-publish@v1.12.4
