### 2025-07-10 cf85b88
fix: add test on time shape
