### 2025-07-10 3a0d32a
fix: add pypi install
