### 2025-07-21 d8ab720
Include area of a pixel in mask_projected_limb
