### 2025-07-21 378281d
Changed mask_projected_limb to take N_or_y as arg
