+++ spotter/core.py
@@ -105,7 +105,7 @@ def equator_coords(phi=None, inc=None, obl=None):
     return jnp.array([x, y, z])
 
 
-def mask_projected_limb(X, N, phase=None, inc=None, u=None, obl=None):
+def mask_projected_limb(N_or_y, phase=None, inc=None, u=None, obl=None):
     """
     Compute mask, projected area, and limb darkening for visible pixels.
 
@@ -131,6 +131,8 @@ def mask_projected_limb(X, N, phase=None, inc=None, u=None, obl=None):
     limb_darkening : ndarray
         Limb darkening factor for each pixel.
     """
+    N, _ = _N_or_Y_to_N_n(N_or_y)
+    X = vec(N)
     d = distance(X, equator_coords(phase, inc, obl))
     mask = d < jnp.pi / 2
     z = jnp.cos(d)
@@ -210,9 +212,7 @@ def design_matrix(N_or_y, phase=None, inc=None, u=None, obl=None):
     matrix : ndarray
         Design matrix.
     """
-    X = vec(N_or_y)
-    N, n = _N_or_Y_to_N_n(N_or_y)
-    mask, projected_area, limb_darkening = mask_projected_limb(X, N, phase, inc, u, obl)
+    mask, projected_area, limb_darkening = mask_projected_limb(N_or_y, phase, inc, u, obl)
     geometry = mask * projected_area
     return limb_darkening * geometry / (geometry * limb_darkening).sum()
 
@@ -349,7 +349,6 @@ def render(
     import matplotlib.pyplot as plt
 
     N, n = _N_or_Y_to_N_n(y)
-    X = vec(N)
 
     phi, theta = hp.pix2ang(N, range(n))
 
@@ -358,7 +357,7 @@ def render(
     else:
         rv = 1
 
-    limb_darkening = mask_projected_limb(X, N, phase, inc, u)[2]
+    limb_darkening = mask_projected_limb(N, phase, inc, u)[2]
     rotated = hp.Rotator(
         rot=[phase, np.pi / 2 - inc or 0.0, obl or 0.0], deg=False
     ).rotate_map_pixel(y * limb_darkening * rv)
@@ -393,13 +392,12 @@ def amplitude(N_or_y, inc=None, u=None, undersampling: int = 3) -> callable:
     """
     N, _ = _N_or_Y_to_N_n(N_or_y)
     resolution = hp.nside2resol(N)
-    X = vec(N)
     hp_resolution = resolution * undersampling
     phases = jnp.arange(0, 2 * jnp.pi, hp_resolution)
 
     mask, projected_area, limb_darkening = jax.jit(
-        jax.vmap(jax.jit(mask_projected_limb), in_axes=(None, None, 0, None, None))
-    )(X, N, phases, inc, u)
+        jax.vmap(jax.jit(mask_projected_limb), in_axes=(None, 0, None, None))
+    )(N, phases, inc, u)
 
     geometry = mask * projected_area
     norm = (geometry * limb_darkening).sum(1)
@@ -568,11 +566,11 @@ def integrated_spectrum(
     """
     spectra = jnp.atleast_2d(spectra)
     if period is None:
-        mask, projected, limb = mask_projected_limb(vec(N), N, 0.0, inc=inc)
+        mask, projected, limb = mask_projected_limb(N, 0.0, inc=inc)
         limb_geometry = projected * mask * limb
         spectra_shifted = spectra.T
     else:
-        mask, projected, limb = mask_projected_limb(vec(N), N, phase, inc=inc)
+        mask, projected, limb = mask_projected_limb(N, phase, inc=inc)
         w_shift = doppler_shift(theta, phi, period, radius, phase)
         dw = wv[1] - wv[0]
         shift = w_shift[:, None] * wv / dw
