### 2025-07-21 0e66ebc
Bump version to 0.0.10
