### 2025-07-30 cb3ead9
Added function to integrate transit spectrum
