+++ spotter/light_curves.py
@@ -32,6 +32,8 @@ def design_matrix(star: Star, time: ArrayLike) -> ArrayLike:
     matrix : ndarray
         Design matrix.
     """
+    import pdb
+    pdb.set_trace()
     if star.u is not None:
         if len(star.y) == 1:
             return jax.vmap(
@@ -135,7 +137,8 @@ def transit_design_matrix(star, x, y, z, r, time=None):
     )
 
     transited_y = utils.sigmoid(distance - r, 1000.0)
-
+    import pdb
+    pdb.set_trace()
     return X * jnp.where(z >= 0, transited_y, jnp.ones_like(transited_y))
 
 
@@ -187,3 +190,4 @@ def transit_light_curve(
         ).T
         * norm
     )
+
