### 2025-08-15 127126b
Reverting back to calc design_matrix with norm
