### 2025-08-19 edf4db7
Better integrated_spectrum, add _check_spectrum
