+++ spotter/core.py
@@ -214,8 +214,7 @@ def design_matrix(N_or_y, phase=None, inc=None, u=None, obl=None):
     """
     mask, projected_area, limb_darkening = mask_projected_limb(N_or_y, phase, inc, u, obl)
     geometry = mask * projected_area
-    return limb_darkening * geometry / (geometry * limb_darkening).sum()
-    # return limb_darkening * geometry
+    return limb_darkening * geometry
 
 
 def flux(y, inc=None, u=None, phase=None, obl=None):
@@ -534,15 +533,15 @@ def shifted_spectra(spectra, shift):
 
 
 def integrated_spectrum(
-    N, theta, phi, period, radius, wv, spectra, phase, inc, normalize=True
+    design_matrix, theta, phi, period, radius, wv, spectra, phase, normalize=True
 ):
     """
     Compute the integrated spectrum of a rotating star.
 
     Parameters
     ----------
-    N : int
-        HEALPix nside.
+    design_matrix : array_like
+        Flux design matrix for a HEALPix map.
     theta : array_like
         Colatitude in radians.
     phi : array_like
@@ -567,22 +566,18 @@ def integrated_spectrum(
     """
     spectra = jnp.atleast_2d(spectra)
     if period is None:
-        mask, projected, limb = mask_projected_limb(N, 0.0, inc=inc)
-        limb_geometry = projected * mask * limb
         spectra_shifted = spectra.T
     else:
-        mask, projected, limb = mask_projected_limb(N, phase, inc=inc)
         w_shift = doppler_shift(theta, phi, period, radius, phase)
         dw = wv[1] - wv[0]
         shift = w_shift[:, None] * wv / dw
-        limb_geometry = projected * mask * limb
         spectra_shifted = shifted_spectra(spectra.T, shift)
 
     if normalize:
         integrated_spec = jnp.sum(
-            spectra_shifted * limb_geometry[:, None], 0
-        ) / jnp.sum(limb_geometry[:, None] * spectra.T)
+            spectra_shifted * design_matrix.T, 0
+        ) / jnp.sum(design_matrix.T * spectra.T)
     else:
-        integrated_spec = jnp.sum(spectra_shifted * limb_geometry[:, None], 0)
+        integrated_spec = jnp.einsum("ij,ij->j", design_matrix.T, spectra_shifted)
 
     return integrated_spec
