### 2025-08-19 72e46f1
Added normalize kwarg to core.design_matrix
