### 2025-08-19 be13d61
fix: light_curves.design_matrix with impl function
