### 2025-08-20 5d2c98b
fix: design_matrix calc in doppler.spectrum
