### 2025-08-22 #8305 @adonath
Thanks @lgrcia, the paper draft looks good! I just have a comment related to the `Precision` section: right now there is nothing the reader can draw from it. I would recommend to add a small quantitative statement along the lines of "we find a typical agreement up to a relative tolerance 0.1% in predicted flux and or spectrum".

Once this is added you can go ahead and:
