### 2025-08-27 dbfa68b
Changing RV sign to match the rotation direction
