### 2025-10-01 37870b1
feat: add text about precision and resolution
