### 2025-10-07 313d5f7
Explicitly account for star inc for doppler shift
