### 2025-11-19 d3f217d
feat: add Vatsal Panwar
