### 2025-12-15 b12b3e2
fix: Vatsal Panwar affiliation
